Split fixtures from gherkin utilities definitions
