+++ tests/e2e/smoke/homepage.spec.ts
@@ -1,4 +1,5 @@
-import { test, type StepDefinition } from '@tests/fixtures'
+import { test } from '@tests/fixtures'
+import type { StepDefinition } from '@tests/fixtures/gherkin'
 import { ui, languageKeys } from '@i18n/ui'
 import { HomePage, userInHome } from '@tests/pages/HomePage'
 import { sectionsConfig } from '@domain/section'
